Introduce Atomic design
