+++ src/components/molecules/InfoWindow/InfoWindow.css
