--- src/containers/Confirmation.js
@@ -1,10 +1,0 @@
-import { connect } from 'react-redux';
-import Confirmation from '../components/Confirmation/Confirmation';
-
-const mapStateToProps = state => ({
-  jwt: state.user.jwt,
-  chargeResponse: state.pay.chargeResponse,
-});
-
-export default connect(mapStateToProps, null)(Confirmation);
-
