+++ src/containers/organisms/Confirmation.js
@@ -1,0 +1,10 @@
+import { connect } from 'react-redux';
+import Confirmation from '../../components/organisms/Confirmation/Confirmation';
+
+const mapStateToProps = state => ({
+  jwt: state.user.jwt,
+  chargeResponse: state.pay.chargeResponse,
+});
+
+export default connect(mapStateToProps, null)(Confirmation);
+
